refactor: views
