+++ src/router/index.js
@@ -5,7 +5,7 @@ import PipelineReportView from '../views/pipeline/ReportView.vue'
 import PipelineConfigsView from '../views/config/ConfigsView.vue';
 import PipelineConfigView from '../views/config/ConfigView.vue';
 import ProfileView from "../views/ProfileView.vue";
-import QuickStartView from "../views/QuickStart.vue";
+import SCMDashboard from "../views/SCMDashboard.vue";
 import { createAuthGuard } from "@auth0/auth0-vue";
 
 const isAuthEnabled = process.env.VUE_APP_AUTH_ENABLED === 'true';
@@ -71,9 +71,9 @@ if (isAuthEnabled) {
     },
     {
       beforeEnter: createAuthGuard(),
-      path: "/quickstart",
-      name: "quickstart",
-      component: QuickStartView
+      path: "/scm/summary",
+      name: "scmSummary",
+      component: SCMDashboard
     }
   ]
 } else {
@@ -122,9 +122,9 @@ if (isAuthEnabled) {
       component: PipelineReportView
     },
     {
-      path: "/quickstart",
-      name: "quickstart",
-      component: QuickStartView
+      path: "/scm/dashboard",
+      name: "scmDashboard",
+      component: SCMDashboard
     }
   ]
